# contracts/api/marketplace-central.openapi.yaml
openapi: 3.1.0
info:
  title: Marketplace Central API
  version: 2026-04-02
paths:
...
  /connectors/vtex/validate-connection:
    post:
      summary: Validate VTEX API credentials
      operationId: validateVTEXConnection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - vtex_account
              properties:
                vtex_account:
                  type: string
                  description: VTEX account name
      responses:
        "200":
          description: Connection successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [connected]
                  vtex_account:
                    type: string
        "400":
          description: Missing vtex_account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "401":
          description: Invalid VTEX credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "502":
          description: VTEX API unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /connectors/melhor-envio/auth/start:
    get:
      summary: Start Melhor Envio OAuth flow
      operationId: startMelhorEnvioOAuth
      responses:
        "302":
          description: Redirect to Melhor Envio authorization page
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorErrorResponse'
        "500":
          description: Failed to create OAuth state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorErrorResponse'
  /connectors/melhor-envio/auth/callback:
    get:
      summary: Handle Melhor Envio OAuth callback
      operationId: handleMelhorEnvioOAuthCallback
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: OAuth exchange successful
        "400":
          description: Missing code or state mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorErrorResponse'
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorErrorResponse'
        "500":
          description: Failed to save token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorErrorResponse'
        "502":
          description: Token exchange failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorErrorResponse'
  /connectors/melhor-envio/status:
    get:
      summary: Check Melhor Envio OAuth status
      operationId: getMelhorEnvioOAuthStatus
      responses:
        "200":
          description: OAuth connection status
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected:
                    type: boolean
        "405":
          description: Method not allowed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorErrorResponse'
        "500":
          description: Failed to read token state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorErrorResponse'
  /connectors/vtex/publish:
    post:
      operationId: publishToVTEX
      summary: Trigger batch publication to VTEX
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishToVTEXRequest'
      responses:
        '201':
          description: Batch created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishToVTEXResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectorErrorResponse'
...
components:
  schemas:
...
    PublishToVTEXRequest:
      type: object
      required:
        - vtex_account
        - products
      properties:
        vtex_account:
          type: string
        products:
          type: array
          items:
            $ref: '#/components/schemas/ProductForPublish'
...
    ProductForPublish:
      type: object
      required:
        - product_id
        - name
        - sku_name
        - ean
        - category
        - brand
        - base_price
      properties:
        product_id:
          type: string
        name:
          type: string
        description:
          type: string
        sku_name:
          type: string
        ean:
          type: string
        category:
          type: string
        brand:
          type: string
        cost:
          type: number
          format: double
        base_price:
          type: number
          format: double
        image_urls:
          type: array
          items:
            type: string
        specs:
          type: object
          additionalProperties:
            type: string
        stock_qty:
          type: integer
        warehouse_id:
          type: string
        trade_policy_id:
          type: string
    PublishToVTEXResponse:
      type: object
      properties:
        batch_id:
          type: string
        total_products:
          type: integer
        validated:
          type: integer
        rejected:
          type: integer
        rejections:
          type: array
          items:
            type: object
            properties:
              product_id:
                type: string
              error_code:
                type: string
...
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: >
                Module-specific error code following the pattern MODULE_ENTITY_REASON.
                Known codes include: CATALOG_PRODUCT_NOT_FOUND, CATALOG_ENRICHMENT_INVALID,
                CLASSIFICATIONS_ENTITY_NOT_FOUND, CLASSIFICATIONS_DUPLICATE_NAME,
                MARKETPLACES_ACCOUNT_INVALID, MARKETPLACES_POLICY_INVALID,
                MARKETPLACES_METHOD_NOT_ALLOWED, MARKETPLACES_FEESCHEDULE_PARAM_MISSING,
                MARKETPLACES_DEFINITIONS_FETCH_FAILED, MARKETPLACES_INTERNAL_ERROR,
                PRICING_SIMULATION_INVALID, PRICING_SIMULATION_INVALID_IDENTITY,
                PRICING_SIMULATION_INVALID_BASE_PRICE, PRICING_SIMULATION_INVALID_VALUES,
                PRICING_SIMULATION_NOT_FOUND, PRICING_METHOD_NOT_ALLOWED,
                CONNECTORS_BATCH_NOT_FOUND, CONNECTORS_RETRY_MISSING_PRODUCT,
                INTERNAL_ERROR, INVALID_REQUEST, NOT_FOUND, CONFLICT.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    ConnectorErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: object
              properties:
                code:
                  type: string
                  enum:
                    - CONNECTORS_PUBLISH_MISSING_NAME
                    - CONNECTORS_PUBLISH_MISSING_SKU
                    - CONNECTORS_PUBLISH_MISSING_CATEGORY
                    - CONNECTORS_PUBLISH_MISSING_BRAND
                    - CONNECTORS_PUBLISH_MISSING_PRICE
                    - CONNECTORS_PUBLISH_MISSING_STOCK
                    - CONNECTORS_PUBLISH_ALREADY_IN_PROGRESS
                    - CONNECTORS_PUBLISH_DEPENDENCY_FAILED
                    - CONNECTORS_VTEX_VALIDATION
                    - CONNECTORS_VTEX_NOT_FOUND
                    - CONNECTORS_VTEX_TRANSIENT
                    - CONNECTORS_VTEX_AUTH
                    - CONNECTORS_METHOD_NOT_ALLOWED
                    - CONNECTORS_ME_METHOD_NOT_ALLOWED
                    - CONNECTORS_ME_STATE_GENERATION_FAILED
                    - CONNECTORS_ME_STATE_MISSING
                    - CONNECTORS_ME_STATE_MISMATCH
                    - CONNECTORS_ME_CODE_MISSING
                    - CONNECTORS_ME_TOKEN_EXCHANGE_FAILED
                    - CONNECTORS_ME_TOKEN_SAVE_FAILED
                    - CONNECTORS_ME_STATUS_STORE_FAILED
                    - CONNECTORS_EXECUTOR_INTERNAL
                    - CONNECTORS_RETRY_MISSING_PRODUCT
                    - CONNECTORS_BATCH_NOT_FOUND